+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## master
+
+- Use background delete propagation when reconciling job, daemon set, and
+  deployment absence
+
 ## v2.0.0
 
 - Ditch github.com/ericchiang/k8s in favor of k8s.io/client-go